fix: fix ui updates
