+++ src/electron.ts
@@ -19,6 +19,7 @@ registerCrashReporter()
 //   const electronTelemetry = electronTelemetryFactory('9b0c8ca04a361eb8160d98c5', buildOptions)
 // }
 
+// disable-dev-shm-usage is required to run the debug console
 app.commandLine.appendSwitch('--no-sandbox --disable-dev-shm-usage')
 app.whenReady().then(() => {
   backendRpc.on(makeOpenDialogRpc(), async request => {
@@ -69,8 +70,6 @@ async function createWindow() {
     }
   })
 
-  console.log('icon path', iconPath)
-  mainWindow.webContents.openDevTools({ mode: 'detach' })
   // Load the index.html of the app.
   if (isDev()) {
     mainWindow.loadURL('http://localhost:8080')
